+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changes in Matrix iOS SDK in 0.11.0 ()
 
 Improvements:
 * MXRestClient: Add Matrix filter API.
+* MXRoom: Add an asynchronous methods for liveTimeline, state and members
 * MXRoomState: Add a membersCount property to store members stats independently from MXRoomMember objects.
 * MXRoomSummary: Add a membersCount property to cache MXRoomState one.
 * MXRoomSummary: Add a membership property to cache MXRoomState one.